Better error handling
When an event returns an error from its callback, it is run through an
error handler (that can be replaced). It then either propagates the
error outside of run* functions, or just stays happy if the error
handler returns Ok(()).

The default error handler simply passes the error through.

Also, the loop refuses to run when empty (which would just block
forever).